+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.1] - 2025-12-12
+
+### Added
+
+- `share` command to share files/folders (--anyone or --email)
+- `unshare` command to remove permissions
+- `permissions` command to list permissions on a file
+
 ## [0.1.0] - 2025-12-12
 
 ### Added